requirements.txt
flask # Please don't make fun of me
Pillow
pywin32 # win32gui replacement
